[Refactor] Add core module for aws funcs, configuration, logging, and asset management; refactor utils to core. Implement S3 upload/download functionality and environment-aware asset path retrieval.
