+++ assignments/movie-sentiment-aws/src/core/logging_config.py
@@ -2,6 +2,9 @@ import logging
 import json
 from logging.handlers import RotatingFileHandler
 from pathlib import Path
+from typing import Dict, Any
+import sys
+from .load_config import config
 
 PROJECT_ROOT = Path(__file__).resolve().parents[2]
 
@@ -11,7 +14,8 @@ class JsonFormatter(logging.Formatter):
     """
     def format(self, record):
         log_record = {
-            "timestamp": self.formatTime(record, self.datefmt)
+            "timestamp": self.formatTime(record, self.datefmt),
+            "pathname": record.pathname
         }
         if isinstance(record.msg, dict):
             log_record.update(record.msg)
@@ -62,3 +66,42 @@ def setup_prediction_logger(config: dict) -> logging.Logger:
         logger.addHandler(logging.NullHandler())
 
     return logger
+
+def setup_logger(config: Dict[str, Any]) -> logging.Logger:
+    """
+    Sets up a rotating file logger and a console logger based on config.
+
+    Args:
+        config (Dict[str, Any]): The application configuration.
+
+    Returns:
+        logging.Logger: The configured logger instance.
+    """
+    logger = logging.getLogger(__name__)
+    logger.setLevel(logging.INFO)
+    formatter = logging.Formatter(
+        "%(asctime)s - %(name)s - %(levelname)s - %(message)s"
+    )
+
+    log_config = config.get("main_logging", {})
+    handler_type = log_config.get("handler")
+
+    if handler_type == "file":
+        log_path_str = log_config.get("path", "assets/logs/app.log")
+        log_path = PROJECT_ROOT / log_path_str
+        log_path.parent.mkdir(parents=True, exist_ok=True)
+
+        # File handler
+        fh = RotatingFileHandler(log_path, maxBytes=5 * 1024 * 1024, backupCount=5)
+        fh.setFormatter(formatter)
+        logger.addHandler(fh)
+
+    # Console handler
+    ch = logging.StreamHandler(sys.stdout)
+    ch.setFormatter(formatter)
+    logger.addHandler(ch)
+
+    return logger
+
+logger = setup_logger(config)
+prediction_logger = setup_prediction_logger(config)
